requirements.txt
Chameleon==2.24
cornice==1.1.0
Mako==1.0.3
MarkupSafe==0.23
nose==1.3.7
PasteDeploy==1.5.2
Pygments==2.0.2
PyMySQL==0.6.7
pyramid==1.5.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
repoze.lru==0.6
simplejson==3.8.1
SQLAlchemy==1.0.10
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
-e git+ssh://56737af189f5cfd896000017@waterlog-pvu.rhcloud.com/~/git/waterlog.git/@674d8f86ddcb92fc71e44fb2eabadfa24b5f131d#egg=waterlog_py-master
WebOb==1.4.1
wheel==0.26.0
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
